+++ src/settings/theme.rs
@@ -43,16 +43,25 @@ pub fn set_theme(app: &mut App, themes: &mut Themes, settings: &mut Settings) {
                 "Player Border Color" => themes.player_border_color = Color::Rgb(r, g, b),
                 "Player Background Color" => themes.player_background_color = Color::Rgb(r, g, b),
                 "Player Highlight Color" => themes.player_highlight_color = Color::Rgb(r, g, b),
+                "Player Inactive Border Color" => {
+                    themes.player_inactive_border_color = Color::Rgb(r, g, b)
+                }
 
                 "Library Border Color" => themes.library_border_color = Color::Rgb(r, g, b),
                 "Library Background Color" => themes.library_background_color = Color::Rgb(r, g, b),
                 "Library Highlight Color" => themes.library_highlight_color = Color::Rgb(r, g, b),
+                "Library Inactive Border Color" => {
+                    themes.library_inactive_border_color = Color::Rgb(r, g, b)
+                }
 
                 "Playlist Border Color" => themes.playlist_border_color = Color::Rgb(r, g, b),
                 "Playlist Background Color" => {
                     themes.playlist_background_color = Color::Rgb(r, g, b)
                 }
                 "Playlist Highlight Color" => themes.playlist_highlight_color = Color::Rgb(r, g, b),
+                "Playlist Inactive Border Color" => {
+                    themes.playlist_inactive_border_color = Color::Rgb(r, g, b)
+                }
 
                 "New Release Border Color" => themes.new_release_border_color = Color::Rgb(r, g, b),
                 "New Release Background Color" => {
@@ -61,16 +70,25 @@ pub fn set_theme(app: &mut App, themes: &mut Themes, settings: &mut Settings) {
                 "New Release Highlight Color" => {
                     themes.new_release_highlight_color = Color::Rgb(r, g, b)
                 }
+                "New Release Inactive Border Color" => {
+                    themes.new_release_inactive_border_color = Color::Rgb(r, g, b)
+                }
 
                 "Main Section Border Color" => themes.main_border_color = Color::Rgb(r, g, b),
                 "Main Section Background Color" => {
                     themes.main_background_color = Color::Rgb(r, g, b)
                 }
                 "Main Section Highlight Color" => themes.main_highlight_color = Color::Rgb(r, g, b),
+                "Main Section Inactive Border Color" => {
+                    themes.main_inactive_border_color = Color::Rgb(r, g, b)
+                }
 
                 "Search Border Color" => themes.search_border_color = Color::Rgb(r, g, b),
                 "Search Background Color" => themes.search_background_color = Color::Rgb(r, g, b),
                 "Search Highlight Color" => themes.search_highlight_color = Color::Rgb(r, g, b),
+                "Search Inactive Border Color" => {
+                    themes.search_inactive_border_color = Color::Rgb(r, g, b)
+                }
 
                 "Help Border Color" => themes.help_border_color = Color::Rgb(r, g, b),
                 "Help Background Color" => themes.help_background_color = Color::Rgb(r, g, b),