+++ src/ui/search/search_album.rs
@@ -24,7 +24,7 @@ pub fn render_searched_album(
         .border_style(if app.searched_album_selected {
             Style::default().fg(theme.main_border_color)
         } else {
-            Style::default()
+            Style::default().fg(theme.main_inactive_border_color)
         })
         .style(Style::default().bg(theme.main_background_color));
 
@@ -35,6 +35,7 @@ pub fn render_searched_album(
         album_block,
         theme.main_highlight_color,
         theme.main_background_color,
+        theme.main_inactive_border_color,
     );
 
     f.render_widget(Clear, content_chunk[1]);