+++ src/ui/search/search_playlist.rs
@@ -22,7 +22,7 @@ pub fn render_searched_playlist(
         .border_style(if app.searched_playlist_selected {
             Style::default().fg(theme.main_border_color)
         } else {
-            Style::default()
+            Style::default().fg(theme.main_inactive_border_color)
         })
         .style(Style::default().bg(theme.main_background_color));
 
@@ -34,6 +34,7 @@ pub fn render_searched_playlist(
         playlist_block,
         theme.main_highlight_color,
         theme.main_background_color,
+        theme.main_inactive_border_color,
     );
 
     f.render_widget(Clear, content_chunk[1]);